firestore connectivity
added the imports to the modules page and uploaded the firebase.js file.

I managed to get the pdf reader working with the old modules page but I had to do it inside of a function.

Will have to test more to see if global variables work the same or if we'll have to implement different strategies to make this work in the list.